requirements.txt
pyTelegramBotAPI==4.20.0
Flask==3.0.0
gunicorn==21.2.0
insightface==0.7.3
onnxruntime==1.16.3
opencv-python==4.8.1.78
Pillow==10.1.0
numpy==1.24.3